fbpreprocessor: clear managed files on logout
If the user logs out, OutputManager needs to clear the list of firmware
dumps that it manages.
However, we do not attempt to delete the files from disk since the
daemon-store could have been unmounted already. The files will be
deleted the next time the user logs in.

BUG=b:328828978
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord
TEST=cros deploy ${DUT} fbpreprocessord
TEST=restart fbpreprocessord
TEST=echo 1 > /sys/kernel/debug/iwlwifi/0000\:00\:14.3/iwlmvm/fw_dbg_collect
TEST=ls -lR /run/daemon-store/fbpreprocessord/ shows a firmware dump in
"processed_dumps"
TEST=dbus-send --system --print-reply=literal
--dest=org.chromium.FbPreprocessor /org/chromium/FbPreprocessor
org.chromium.FbPreprocessor.GetDebugDumps returns the expected output
https://paste.googleplex.com/5033126960562176
TEST=logout, observe that dbus-send --system --print-reply=literal
--dest=org.chromium.FbPreprocessor /org/chromium/FbPreprocessor
org.chromium.FbPreprocessor.GetDebugDumps returns an empty output/
TEST=log back in, firmware dump has been deleted and GetDebugDumps
returns an empty output. Generate a new firmware dump, behavior is as
expected, as described in previous lines.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5357735
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>

+++ fbpreprocessor/output_manager.cc
@@ -77,6 +77,11 @@ void OutputManager::OnUserLoggedIn(const std::string& user_dir) {
 
 void OutputManager::OnUserLoggedOut() {
   LOG(INFO) << "User logged out.";
+  // When we're notified that the user has logged out the daemon-store could be
+  // unmounted already, so we should not try to delete the files from disk.
+  // We merely clear the list of files that we're managing. The files themselves
+  // will be deleted the next time the user logs in.
+  ClearManagedFiles(/*delete_files=*/false);
   user_root_dir_.clear();
 }
 
@@ -181,10 +186,13 @@ void OutputManager::OnExpiredFile() {
   files_lock_.Release();
 }
 
-void OutputManager::DeleteAllManagedFiles() {
+void OutputManager::ClearManagedFiles(bool delete_files) {
   files_lock_.Acquire();
-  for (auto f : files_) {
-    DeleteFirmwareDump(f.fw_dump(), __func__);
+  expiration_timer_.Stop();
+  if (delete_files) {
+    for (auto f : files_) {
+      DeleteFirmwareDump(f.fw_dump(), __func__);
+    }
   }
   files_.clear();
   files_lock_.Release();
@@ -192,7 +200,7 @@ void OutputManager::DeleteAllManagedFiles() {
 
 void OutputManager::DeleteAllFiles() {
   VLOG(kLocalDebugVerbosity) << __func__;
-  DeleteAllManagedFiles();
+  ClearManagedFiles(/*delete_files=*/true);
   base::FileEnumerator files(user_root_dir_.Append(kProcessedDirectory),
                              false /* recursive */,
                              base::FileEnumerator::FILES);
